requirements.txt
langchain==0.0.197
langchainplus-sdk==0.0.8
openai==0.27.8
tiktoken==0.4.0
chromadb==0.3.26
streamlit==1.23.1
pypdf==3.10.0
python-dotenv==1.0.0